Outputs JSON format only for now
issue#6 reports that even JSON formatting wasn't being
output by the CLI in the 'summary' set of commands.
This fixes it to output only JSON format.
Adds a test to validate the formatting of some other options.
A better test is to verify that the values are appearing, without
bothering with checking the precise format (it makes these tests
brittle, since the library might change formatting slightly from version
to version).